# Optimize std.base64.standard.Encoder.encode