Fixes #16291

Improves performance by about 2x, which is better than the "plain" codec of libbase64.

```console
> poop ./repro-libbase64 ./repro-libbase64-plain ./repro-zig-std-old ./repro-zig-std-new

Benchmark 1 (376 runs): ./repro-libbase64
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          13.1ms ± 2.30ms    10.2ms … 23.3ms          7 ( 2%)        0%
  peak_rss           47.9MB ± 56.8KB    47.8MB … 48.0MB          0 ( 0%)        0%
  cpu_cycles         36.7M  ± 5.21M     31.0M  … 62.2M           9 ( 2%)        0%
  instructions       67.6M  ± 6.05      67.6M  … 67.6M          16 ( 4%)        0%
  cache_references    588K  ± 8.96K      569K  …  647K           9 ( 2%)        0%
  cache_misses        229K  ± 32.0K      183K  …  318K           0 ( 0%)        0%
  branch_misses      2.24K  ± 68.0      1.61K  … 2.52K          27 ( 7%)        0%
Benchmark 2 (187 runs): ./repro-libbase64-plain
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          26.7ms ± 3.53ms    22.4ms … 50.4ms         13 ( 7%)        💩+103.6% ±  3.7%
  peak_rss           47.9MB ± 63.2KB    47.8MB … 48.0MB          0 ( 0%)          -  0.0% ±  0.0%
  cpu_cycles         94.6M  ± 9.06M     87.1M  …  166M          16 ( 9%)        💩+158.0% ±  3.2%
  instructions        288M  ± 14.1       288M  …  288M          16 ( 9%)        💩+325.6% ±  0.0%
  cache_references    592K  ± 16.3K      568K  …  663K          10 ( 5%)          +  0.8% ±  0.4%
  cache_misses        193K  ± 35.0K      142K  …  360K           1 ( 1%)        ⚡- 15.6% ±  2.5%
  branch_misses      2.25K  ± 67.3      1.72K  … 2.45K          10 ( 5%)          +  0.7% ±  0.5%
Benchmark 3 (112 runs): ./repro-zig-std-old
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          44.7ms ± 2.85ms    39.9ms … 57.0ms          3 ( 3%)        💩+241.1% ±  3.9%
  peak_rss           46.5MB ± 2.52KB    46.5MB … 46.5MB          0 ( 0%)        ⚡-  2.9% ±  0.0%
  cpu_cycles          170M  ± 5.53M      164M  …  193M           7 ( 6%)        💩+363.1% ±  3.0%
  instructions        729M  ± 11.2       729M  …  729M           8 ( 7%)        💩+978.7% ±  0.0%
  cache_references    590K  ± 12.4K      564K  …  651K           2 ( 2%)          +  0.4% ±  0.4%
  cache_misses        172K  ± 29.8K      134K  …  236K           0 ( 0%)        ⚡- 25.0% ±  2.9%
  branch_misses       334   ± 81.0       220   …  713            2 ( 2%)        ⚡- 85.1% ±  0.7%
Benchmark 4 (212 runs): ./repro-zig-std-new
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          23.6ms ± 3.31ms    17.9ms … 41.3ms          4 ( 2%)        💩+ 80.0% ±  3.5%
  peak_rss           46.5MB ± 2.52KB    46.5MB … 46.5MB          0 ( 0%)        ⚡-  2.8% ±  0.0%
  cpu_cycles         67.4M  ± 2.94M     64.4M  … 82.3M          12 ( 6%)        💩+ 83.8% ±  2.1%
  instructions       80.8M  ± 10.3      80.8M  … 80.8M          85 (40%)        💩+ 19.5% ±  0.0%
  cache_references    557K  ± 21.3K      519K  …  649K           6 ( 3%)        ⚡-  5.2% ±  0.4%
  cache_misses        202K  ± 40.7K      151K  …  390K          10 ( 5%)        ⚡- 11.6% ±  2.6%
  branch_misses      35.9   ± 7.82         6   …   54           39 (18%)        ⚡- 98.4% ±  0.4%
```